Fix -Wdiscarded-qualifiers, -Wredundant-decls, and const-correctness warnings
- Change char* to const char* in function parameters and local variables
  where strings are not modified (libcommon, libsrc, libsrc2, volume_io)
- Remove redundant declarations between minc.h and minc_error.h
- Remove duplicate #include in minc_private.h
- Remove archaic pre-ANSI Sun compiler workaround in minc_useful.h
- Update all corresponding header declarations to match
- Mark unused parameters with (void)param casts

+++ libcommon/ParseArgv.c
@@ -246,7 +246,7 @@ ParseArgv(int *argcPtr, char **argv, ArgvInfo *argTable, int flags)
                char *endPtr;
 
                *(((int *) infoPtr->dst)+i) =
-                  ParseLong(argv[srcIndex], &endPtr);
+                  (int) ParseLong(argv[srcIndex], &endPtr);
                if ((endPtr == argv[srcIndex]) || (*endPtr != 0)) {
                   FPRINTF(stderr,
                   "expected integer argument for \"%s\" but got \"%s\"",
@@ -444,7 +444,7 @@ PrintUsage(ArgvInfo *argTable, int flags)
          FPRINTF(stderr, "\n %s:", infoPtr->key);
          /* Write out padding after the key, followed by the help text.
           */
-         numSpaces = width + 1 - strlen(infoPtr->key);
+         numSpaces = (int) (width + 1 - strlen(infoPtr->key));
          FPRINTF(stderr, "%*s %s", numSpaces, "", infoPtr->help);
          switch (infoPtr->type) {
          case ARGV_INT: {
